pci: Introduce pci_register_io_range() helper function.
Some architectures do not share x86 simple view of the PCI I/O space
and instead use a range of addresses that map to bus addresses. For
some architectures these ranges will be expressed by OF bindings
in a device tree file.

Introduce a pci_register_io_range() helper function that can be used
by the architecture code to keep track of the I/O ranges described by the
PCI bindings. If the PCI_IOBASE macro is not defined that signals
lack of support for PCI and we return an error.

Change-Id: I0cbb40de978292e6efe44509fba2518ef5828d23
Signed-off-by: Liviu Dudau <Liviu.Dudau@xxxxxxx>
Acked-by: Grant Likely <grant.likely@xxxxxxxxxx>
Tested-by: Tanmay Inamdar <tinamdar@xxxxxxx>
Patch-mainline: lkml @ 03/14/14, 15:34:27
Signed-off-by: Tony Truong <truong@codeaurora.org>