pci: OF: Fix the conversion of IO ranges into IO resources.
The ranges property for a host bridge controller in DT describes
the mapping between the PCI bus address and the CPU physical address.
The resources framework however expects that the IO resources start
at a pseudo "port" address 0 (zero) and have a maximum size of IO_SPACE_LIMIT.
The conversion from pci ranges to resources failed to take that into account.

In the process move the function into drivers/of/address.c as it now
depends on pci_address_to_pio() code and make it return an error message.

Change-Id: Ic4fc51f3ff06719493a5f7d55ce981c03eb17b92
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Tested-by: Tanmay Inamdar <tinamdar@apm.com>
Patch-mainline: lkml @ 03/14/14, 15:34:28
[truong@codeaurora.org: resolved minor merge conflict]
Signed-off-by: Tony Truong <truong@codeaurora.org>