pci: Create pci_host_bridge before its associated bus in pci_create_root_bus.
Before commit 7b5436635800 the pci_host_bridge was created before the root bus.
As that commit has added a needless dependency on the bus for pci_alloc_host_bridge()
the creation order has been changed for no good reason. Revert the order of
creation as we are going to depend on the pci_host_bridge structure to retrieve the
domain number of the root bus.

Change-Id: I9d6c97cc1bdd666ca0250abb3a54bda7dc2c868f
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Acked-by: Grant Likely <grant.likely@linaro.org>
Tested-by: Tanmay Inamdar <tinamdar@apm.com>
Patch-mainline: lkml @ 03/14/14, 15:34:29
Signed-off-by: Tony Truong <truong@codeaurora.org>