pci: Introduce a domain number for pci_host_bridge.
Make it easier to discover the domain number of a bus by storing
the number in pci_host_bridge for the root bus. Several architectures
have their own way of storing this information, so it makes sense
to try to unify the code. While at this, add a new function that
creates a root bus in a given domain and make pci_create_root_bus()
a wrapper around this function.

Change-Id: If968407260beb19c8dc683770c369a67224126e3
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Tested-by: Tanmay Inamdar <tinamdar@apm.com>
Patch-mainline: lkml @ 03/14/14, 15:34:30
Signed-off-by: Tony Truong <truong@codeaurora.org>