drivers: usb: Remove BAM reset for HSIC BAM on connect
In host BAM2BAM connect sequence, BAM reset is not required
and can create problems, such as in ODU fast Ethernet cable
connect/disconnect stress test. This might cause the controller
to freeze, since it might be in the middle of something when the BAM
is reset, and this stucks everything, including direct access on
endpoint 0.
WLAN does require such reset so we move this reset to be
close to the hbm_init, where it's safe to reset the BAM.

CRs-fixed: 613557
Change-Id: I3f403dd0b2aa66717a0a4fa72872a595900fa412
Signed-off-by: Lena Salman <esalman@codeaurora.org>