usb: phy-msm-qmp: Move PHY settings into const data
Currently the PHY settings values are hardcoded in the init()
routine.  In order to support different sets of values based
on revision ID or other HW variation, describe the register
offsets and the values to be written in a pairwise table.

In addition, add support for revision 1 devices, which have
different offsets for the QSERDES_COM_* registers.

Change-Id: I5e95ddabd11f4d185f5dbd3443961cfbfb9f258c
Signed-off-by: Jack Pham <jackp@codeaurora.org>