cpufreq: interactive: restructure CPUFREQ_GOV_LIMITS
The cpufreq_interactive_timer gets cancelled and rescheduled
whenever the cpufreq_policy is changed. When the cpufreq policy is
changed at a rate faster than the sampling_rate of the interactive
governor, then the governor misses to change the target frequency
for long duration. The patch removes the need of cancelling the
timers when policy->min is changed.

Change-Id: If78b0a9b3ee42980597bac041dfc66662e6cd307
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
Git-commit: 535a553fc1c4b4c3627c73214ade6326615a7463
Git-repo: https://android.googlesource.com/kernel/common.git
Signed-off-by: Junjie Wu <junjiew@codeaurora.org>