msm: board-8064: Enable touch screen on adp-m platfrom
Enable the touch screen on adp-m platforms. Use platform
data to inform the driver how the reportid provided by
the hardware should be used.
Some I2C devices are not initialized for adp-m platform.
Currently the adp-m and adp platform have same platform type
which causes the adp-m platform to be wrongly detected as
adp platform. Make sure that the adp-m platform is checked
before adp so that the correct values are initialized.

Change-Id: I6ba648bf7b6d597a49ab7351394273377b262557
Signed-off-by: Joveen Joseph Thomas <joveenthomas@codeaurora.org>