qcom: scm: Implement a retry mechanism
There are times when the secure world may be processing
requests that require one particular client on the HLOS
to invoke an scm_call. While waiting for this particular
scm_call, no other scm requests can be processed and the
secure world will return an error code indicating this
situation.

When this error code is encountered, release the scm_call
mutex, wait for a specified amount of time and retry the
scm_call. As much as possible, avoid overhead in the first
scm_call attempt.

Change-Id: I6edce93bb70620fdffcb3ef3fb639cf5b928845b
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
Signed-off-by: Srinivasarao P <spathi@codeaurora.org>