USB: phy-msm-usb: Enable PHY_RETENTION for VDD_MIN on MSM8916
If PHY_RETENTIION is not enable, phy driver does not vote
for vdd_min. As part of host mode enablement, ext_id_irq
check is not added so that PHY_RETENTION can be enabled.
As part of this, target is not entering into VDD_MIN.

To fix this issue, enable PHY_RETENTIOIN if gpio 110
for detecting ID is enabled.

CRs-Fixed: 647209
Change-Id: I66264bdaa17f907efe93eef192470fe0a66dbe6d
Signed-off-by: Sujeet Kumar <ksujeet@codeaurora.org>