usb: dwc3: Add support to configure Suspend Event
Currently Suspend event is enabled by default which is not
required where USB bus suspend/resume support is not used.
Hence add support which allows to enable this event based
on requirement. With this change, Suspend event is disabled
by default. This change avoids issue on receiving suspend
event on devices where it is not supported.

CRs-fixed: 649280
Change-Id: I7364b679af2dbcf2e7f6f902e87113f4eb5ed67c
Signed-off-by: Guang Jian Zeng <gzeng@codeaurora.org>