power: qpnp-charger: enable vtrkl comparator and disable trickle clamp.
An issue with digital logic can cause the charger to get stuck
in trickle charging. In order to avoid this the bootloader
overrides the trickle charge comparator when VBAT > 3.7 V.

Once kernel is initialized, these comparators need to be
re-enabled and the trickle clamp to be disabled.

CRs-Fixed: 646222
Signed-off-by: David Keitel <dkeitel@codeaurora.org>
(cherry picked from commit 79e08321531503297bca589e05a5c8634422b926)

Change-Id: Ida3f18519137e3a7ccfe483cc3168e92d484b429
Signed-off-by: Akhila Musunuri <makhila@codeaurora.org>