arm64: mm: Add tlbi workaround for msm8994
MSM8994 v1 revision has a bug in which the bits [39:38]
of VA are tied to zero. ARM64 Linux uses these
bits for kernel and userspace maps. Due to this bug
it is possible that other processors don't see the
effect of tlb invalidation by VA or ASID and could result
in an abort or data corruption.

The workaround here is to replace the tlb invalidation
by VA and ASID with tlbiall in arm64 kernel.
This workaround is controlled through CONFIG_ARCH_MSM8994_V1_TLBI_WA.

Change-Id: Ie36c168881e61007d5d5383d989d7ee6adbd1d92
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>