input: synaptics: secure touch support
Add secure touch support to the synaptics RMI4 and DSX drivers.
Expose the new interface through sysfs, to enable/disable
secure touch, and to wait for touch events.

When secure touch is active, interrupts are forwarded via sysfs
to a listening layer in userspace and processed in TrustZone.
Secure touch is aborted on power events.

Change-Id: If4c6b76d020ef061338378a8737702cec7bada33
Acked-by: Christian Bolis <cbolis@qti.qualcomm.com>
Signed-off-by: Keith Fallows <keithf@codeaurora.org>
[mattw@codeaurora.org: resolved trivial Kconfig conflict]
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>