mtd: msm_qpic_nand: fix a race condition in page read
During a recent performance optimization patch, there was race condition
introduced for a page read opeartion. If there are more than one client
for a NAND read, then its possible to hit this bug.

During the race condition, the READ_LOCATION_0 is not written and this
leads to reading an incorrect number of bytes per codeword sometimes. Fix
this protecting the code with a mutex.

CRs-Fixed: 650682
Change-Id: I58426a2b9005619901173c3e87da7169024e5041
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>