net: usb: Add runtime decision whether to support ODU
The path used for ODU has a different flow in usbnet than
SW paths, because we need to configure the HW path.
Since SMSC device needs to be used for both
usecases - SW and HW(via IPA), we need to add support
to ask in runtime what support is configured for current run.

CRs-fixed: 600475
Change-Id: Ic06d8eec3b3a0cd0960a21274902cb9a3b84be9a
Signed-off-by: Lena Salman <esalman@codeaurora.org>