cpufreq: interactive: prevents the frequency to directly raise above the hispeed_freq from a lower frequency.
When the load was below go_hispeed_load, there is a possibility that
choose_freq() would return a frequency which would be higher than the
hispeed_freq. According to the policy we should first jump to the
hispeed_freq, stay there for above_hispeed_delay and then be allowed to
raise higher than that.

Added a check to prevent the frequency to be directly raised to
something higher than the hispeed_freq.

Change-Id: I74ec5468e7ce0b56b9a9fae6a1930e52d53df3bf
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Git-commit: 7f3de5191eeca13d7daad972aedddc245073cdbf
Git-repo: https://android.googlesource.com/kernel/common.git
Signed-off-by: Junjie Wu <junjiew@codeaurora.org>