ARM: dts: msm: move GICv3 to one contiguous rdist region on thulium
The GICv3 spec indicates that the GICR_TYPER[4] specifies the
last redistributor region in a group of contiguous rdist pages.
It so happens that thulium's rdist regions are physically
contiguous and designed to have only the last rdist region set
GICR_TYPER[4]. Previous bugs in the hardware prevented this from
working properly, but recent changes have made this
feasible.

Without this change we'll now end up mapping each region
separately, which implies that there's no guarantee of virtually
contiguous mappings. Thus when scanning for contiguous regions,
we'll note that the first region does not have GICR_TYPER[4] set,
and then dereference an illegal page of virtual memory when
trying to check the second.

Long story short, we need this change to not crash while booting.

Change-Id: I172616aa69838c35d7741fc82e83f93f6f7fb458
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>