Revert "defconfig: Enable config to allow vmalloc savings"
This reverts commit c69c0451388c7ba3ca7bde8c3b6cddfefcba54c1.

CONFIG_VMALLOC_SAVING flag enables us to use lowmem
mappings which generally goes unused because of
non-hlos reservations present in lowmem region. But
after enabling the flag as the Vmalloc region of
240 MB gets split into two regions of 156 MB and
92 MB, PIL driver is not able to get vmalloc
region of 84 MB needed for loading modem heap as
single contiguous region of 84 MB is not available
because of vmalloc region fragmentation.

Change-Id: Id9647a4c5f2041142f1ea2e59d601863ec28bb5d
Signed-off-by: Panwar Vivek <vpanwa@codeaurora.org>