usb: dwc3-msm: Fix resume sequence when cable is reconnected
The test for 'mdwc->vbus_active' was incorrectly skipping
the necessary HS PHY reset step when resuming the controller
upon cable reconnect, since it would be set to true. Instead
this should be performed regardless of cable connected status
and rather by whether the controller is coming out of a
power-collapsed state. This can be combined with the existing
post-clocks power collapse exit sequence, which also already
does a UTMI interface reset.

While at it, move the SSPHY resume to after turning back on
the reference clock as it can be done earlier as per the HW
recommended sequence and eliminates an unncessary local flag.

Change-Id: Id02c5620d422ad28ae720af83dd5d02b2b796c31
Signed-off-by: Jack Pham <jackp@codeaurora.org>