i2c-msm-v2: control clocks irqs and bam per transfer
Move control of clocks and irqs from runtime-pm to transfer
begin and end as clk and irq are local and there is no penalty
in doing per xfer. But Bus scaling is expensive as has a 1 second
delay and thus done per PM which is less often. Also initialize bam
when clocks are turned on only. This is required when the i2c
core is accessed from other execution enviornments.

Change-Id: If372a94394d10cc8040d122c21e839a28fbaae9a
Signed-off-by: Gilad Avidov <gavidov@codeaurora.org>
Signed-off-by: Ankit Gupta <ankgupta@codeaurora.org>