soc: qcom: Add system health monitor
System Health Monitor (SHM) passively monitors the health of the
peripherals connected to the application processor. Software components
in the application processor that experience communication failure can
request the SHM to perform a system-wide health check. If any failures
are detected during the health-check, then a subsystem restart will be
triggered for the failed subsystem.

Change-Id: If9913e4489f06c74e40443bcc673ded0336a26af
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>