ARM: dts: msm: add vddio supply entry to dsi controller power module
The vddio supply is used by both the DSI panel as well as the DSI
controller. Currently, this supply is only listed as part of the panel
power module. Add this supply to the controller module as well. This
ensures correct behavior if any of the power modules are selectively
controlled.

CRs-Fixed: 683597
Change-Id: I4e2b31c8aea293dc1ce81f5b616cc687c871592f
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>