usb: phy: msm-hsphy: Fix host bus suspend/resume sequeunce
Fix power collapse issue in the multiport PHY: when USB port1 or USB
port 0 are disconnected, port 2-3 doesn't respond to plug in event and
doesn't do enumeration. it is caused by the power collapse when USB
core enter to suspend state.

It seems that Async interrupt is not being generated on connecting
device while host is in bus suspend state.

Change-Id: Ia9727be053bf427bd64bfef7d96002d68eafead1
Signed-off-by: Tanya Finkel <tfinkel@codeaurora.org>
Signed-off-by: Mayank Rana <mrana@codeaurora.org>