msm: pm: Fix error in choosing suspend warmboot path
The last core that finishes the suspend ops is always online and that
flag can be used to determine if the CPU needs to enter through hotplug
warmboot path or not.  A hotplugged core is marked offine before its
called into its terminating function.  This results in a incorrect
num_online_cpus() for the last but one core going down and could result
in hotplugged core entering via suspend path.

Change-Id: I9655040ab0099b12dc1b7a84ea2d865e9da77d68
Signed-off-by: Panwar Vivek <vpanwa@codeaurora.org>