usb: tty: Add more buffer space and make the buffer use direct
Due to stall in high throughput in DUN+SoftAP mode, we
needed to improve the way we transfer the buffers.
Here we enlarge the buffer space so we have more space
to buffer the high throughput data and switch to direct
buffer use, excluding the need and the overhead
of a work queue.

CRs-fixed: 681446
Change-Id: I77e03d8317803b0ddbb98f1575e5b0e73b1e20b4
Signed-off-by: Lena Salman <esalman@codeaurora.org>
[jackp@codeaurora.org: modified to use tty_buffer_set_limit]
Signed-off-by: Jack Pham <jackp@codeaurora.org>