usb: dwc3-msm: Fix LPM race condition on cable disconnect
In some cases, the pm_runtime_suspend may be called by the usb_bam.c
when there is pending lpm flag set after waiting IPA resource manager
to release the consumer resource. However, if this is done when the
usb cable was disconnected and otg state has not yet changed to IDLE,
then it means otg state machine is running and we race against it.
So cancel the LPM for now, and the otg state machine will go to LPM
later, after completing transition to IDLE state.

Change-Id: Iaf2d8bd3156245c83a1b9bab4b9d62c82f16a7c2
CRs-Fixed: 687668
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>