msm: cpp: Validate frame message before manipulating it.
CPP frame message is used to send all frame data
to Microcontroller. It is sent every frame. CPP kernel
driver has to add information to it before transfer it.
The message has to be validated before manipulations.
If it is not valid the message and corresponding frame
are discarded.
Signed-off-by: Manish Poddar <mpoddar@codeaurora.org>
Signed-off-by: Natiq Ahmed <mnatiq@codeaurora.org>
Conflicts:
	drivers/media/platform/msm/camera_v2/pproc/cpp/msm_cpp.c
	drivers/media/platform/msm/camera_v2/pproc/cpp/msm_cpp.h
	include/media/msmb_pproc.h
Change-Id: Ib5b9b5d2e1886d3d671966b693ce212d58e34041