iommu/arm-smmu: allow translation stage to be forced on the cmdline
When debugging and testing code on an SMMU that supports nested
translation, it can be useful to restrict the driver to a particular
stage of translation.

This patch adds a module parameter to the ARM SMMU driver to allow this
by restricting the ability of the probe() code to detect support for
only the specified stage.

Change-Id: Ic2aea66afd9f9973f8809cc3a05f0104e875dd87
Signed-off-by: Will Deacon <will.deacon@arm.com>
Git-commit: 4cf740b0b6628bda1e5c9201ae0d4f56fc6c06a5
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
[mitchelh: context fixups]
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>