ARM: dts: msm: add the default value for VMBMS S2 FIFO count
The register value of VMBMS S2 FIFO count is changed according
to the SOC level, it is 5 by default, and will be changed to 2
when the SOC is lower than low_soc_threshold. This value is not
reset across a warm reset.

If the device resets when SOC lower than low_soc_threshold, the
VMBMS S2 FIFO count will always be 2 even when the SOC is higer
than low_soc_threshold, it should be 5 in this case.

Set a default value 5 for VMBMS S2 FIFO count when device boots up.

Change-Id: I49a3e50dd2a5f13d2bcb116497a62dc71bcc4dcb
Signed-off-by: Jie Cheng <rockiec@codeaurora.org>