scsi: ufs-msm: probe UFS only if it is the boot device
Boot device can be either UFS or eMMC which means if eMMC is the boot
device, probing UFS device is not desirable as it's not going to be
used after probing. Kernel command line parameter "androidboot.bootdevice"
tells the kernel about the boot device so look at this boot device
parameter to know whether to probe UFS device or not.

Change-Id: I4f2b4ce56276658e4c72790b889d5a38ec78e518
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>