usb: gadget: Resolve recursive spin locking during write to IPA on suspend
When the USB core needs write new data to IPA, it locks the port lock
spin lock in the bam_data_write_toipa() function and calls ep_queue().
If this is done during USB bus suspend state, it triggers resume event
handling in which the port lock spin lock is locked again, and this leads
to a crash. This patch fixes the issue by un-locking the lock before the
ep_queue() call.

CRs-fixed: 698859
Change-Id: I0e813a3036e5c401dc00225fb14761bef15b6832
Signed-off-by: Danny Segal <dsegal@codeaurora.org>