i2c-msm-v2: remove full bam init/teardown per transfer
The clocks are turned on and off per transfer. After the
clocks are off and before using bam for next transfers,
the bam pipes need to be disconnected and connected back,
but full bam teardown and initialization is not required.
Remove the redundant bam flushing in a case of error
on bam tarnsfer as the disconnection and connection of the
pipes effectivly flashing the bam pipes every transfer.

Change-Id: I275287ebd3ef969e93bd00d222cba56e9c887e7c
Signed-off-by: Kiran Gunda <kgunda@codeaurora.org>
Signed-off-by: Sana Venkat Raju <c_vsana@codeaurora.org>
Signed-off-by: Gilad Avidov <gavidov@codeaurora.org>
Signed-off-by: Ankit Gupta <ankgupta@codeaurora.org>
Signed-off-by: Satish Kamuju <skamuj@codeaurora.org>