regulator: Snapshot cpr-regulator driver
Pick up the cpr-regulator driver as of msm-3.10 commit
acdce027 (Merge "defconfig: arm64: Enable ONESHOT_SYNC for
msm8994").

We drop two names in vdd_apc_name[] because it's an 8926
workaround (we're not going to support that SoC here) and
regulator_get() nowadays will insert a dummy regulator if the
device is in DT and the supply isn't specified.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>