of_batterydata: add qcom,batt-id-range-pct support
The optional qcom,batt-id-range-pct property will
be used to determine a valid range for a given qcom,batt-id-kohm.

This option puts a boundary on the closest match found by
of_batterydata_get_best_profile.

The advantage of this is to prevent loading battery profiles
based on battery ID resistance values which are too far off from
qcom,batt-id-kohm and thus leading to undesirable configuration
for battery related properties.

Change-Id: I9d27550c73ef626f08428d05a23354e6fd91eb20
Signed-off-by: David Keitel <dkeitel@codeaurora.org>