power: smb1360: Fix the shutdown logic for SMB1360
Force the part in/out of shutdown irrespective of the
previous state.

While at it, do not use a DT based check to bring the
part out of shutdown. This is for devices which are
configured for shutdown by default.

Change-Id: Ib74f06e6471ad1ec0e10d5875e27555b6c499788
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>
(cherry picked from commit e21731f29ad7e63fd171dc383c419d12d9f2d2d0)