msm: rpm-regulator-smd: add support to send default requests
Add support for a boolean device tree property which can be
specified in order to cause default initial parameter values to
be sent to the RPM before any consumers have made their own
requests.  This feature can be used in situations where it is
not desirable to enable a RPM controlled regulator but where some
other SMD parameter must be sent.

Change-Id: Ic9ed813064b0d3e3884fc328af7cb6b8562b9dfe
Signed-off-by: David Collins <collinsd@codeaurora.org>