msm: camera: isp: Update burst count correctly in overflow case
During overflow recovery burst count value needs to be restored.
It can be updated directly without a call to the framedrop API.
This simplifies code logic and avoids unnecessay initializations
needed when using the API. Also when the burst count is computed,
the hardware needs to be configured with the updated value.

CRs-fixed: 724338
Change-Id: Ic55729752fd8251d2c253c316f4c38454c497542
Signed-off-by: Venu Yeshala <vyeshala@codeaurora.org>