cpufreq: interactive: Limit timer rearm to just once on limits change
Currently when a CPU changes its limits, its timer is canceled and
then rearmed from cpufreq_governor_interactive(). However this can
lead to timer not getting to run for a long duration, for example
due to rapid changes in scaling_min.
This patch ensures that the cancelation and re-arming of timer is
not done more than once before the timer function gets a chance to
run.

Change-Id: I971ff00eae05481e3296cd2f8ee42d23a527c197
Signed-off-by: Rohit Gupta <rohgup@codeaurora.org>