drivers: usb: Stop transfers on MSG_HALT
Some of the versions of windows XP send MSG_HALT
without sending RNDIS_OID_GEN_CURRENT_PACKET_FILTER
before. This causes the interface to lose IP upon consequent
enable.
With this fix we close the transfer with the IPA correctly, and
allowing the next connect to open the connection correctly.

CRs-fixed: 711489
Change-Id: Iaf50803853581ff5bb1df4b3407e5614797defd1
Signed-off-by: Lena Salman <esalman@codeaurora.org>