cpufreq: cpu-boost: Restrict migration syncs only to unrelated CPUs
Currently we do not consider the relationship between CPUs while
performing migration syncs. So if the boost sync feature is turned
ON from userpspace we could end up performing syncs for task
migrations between the CPUs which share the same policy.
To avoid this restrict the sync operation for task migrations only
across CPUs which do not share the same policy.

The above is restricted to only when task migrates from little to
big.

Change-Id: Icd00b97663bf95b9d1fea568737ad123d1249ac7
Signed-off-by: Rohit Gupta <rohgup@codeaurora.org>
Signed-off-by: Hanumath Prasad <hpprasad@codeaurora.org>