msm: kgsl: reduce maximum inflight for multiple active contexts
Modify the maximum allowable inflight depending on the number of
actively rendering contexts. An active context is one that has
submitted a command within the last 500 milliseconds.

If more than one active context is rendering drop the inflight
to a lower threshold (default 4) to reduce the number of things
sitting in the ringbuffer - this should improve cross-context
latency.

Change-Id: Ic0dedbad277bb9a2d9f0198d7fb3b52308b32124
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>