regulator: core: Add REGULATOR_EVENT_PRE_VOLTAGE_CHANGE (and ABORT)
In some cases we need to know when a regulator is about to be changed.
Add a way for clients to be notified.  Note that for set_voltage() we
don't necessarily know what voltage we'll end up with, so we tell the
client what the range will be so they can prepare.

Change-Id: Ibdaddc1e86dcf4d1db4f9e9e48cc2328c70d5def
Signed-off-by: Heiko Stübner <heiko@sntech.de>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Mark Brown <broonie+linaro@kernel.org>
Git-commit: 7179569aeb52197fd2a9909ba226c4c9cc0e2e2a
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
[collinsd@codeaurora.org: resolve merge conflicts]
Signed-off-by: David Collins <collinsd@codeaurora.org>